move apply-only types back to apply.go
